FIT image conversion script and training scripts
FIT image conversion script and training scripts for images from RMIT's ROO telescope. Does not include labeling of ROO images as that was done manually. Python training script was done for debugging on local machine were training of models was done on Google Colab using the code blocks in the Jupyter Notebook file.